+++ src/gocept/selenium/webdriver.py
@@ -12,9 +12,7 @@
 #
 ##############################################################################
 
-from selenium.webdriver.firefox.firefox_binary import FirefoxBinary
 from selenium.webdriver.firefox.firefox_profile import FirefoxProfile
-import ast
 import atexit
 import gocept.selenium.wd_selenese
 import os
@@ -23,7 +21,6 @@ import selenium.webdriver
 import shutil
 import sys
 import tempfile
-import urllib2
 
 
 # work around Python 2.4 lack of absolute_import,
@@ -33,16 +30,7 @@ plonetesting = __import__('plone.testing', {}, {}, [''])
 
 class Layer(plonetesting.Layer):
 
-    # hostname and port of the Selenium RC server
-    _server = os.environ.get('GOCEPT_SELENIUM_SERVER_HOST', 'localhost')
-    _port = int(os.environ.get('GOCEPT_SELENIUM_SERVER_PORT', 4444))
-
     _browser = os.environ.get('GOCEPT_WEBDRIVER_BROWSER', 'firefox')
-
-    @property
-    def is_remote(self):
-        return ast.literal_eval(
-            os.environ.get('GOCEPT_WEBDRIVER_REMOTE', 'True'))
 
     def setUp(self):
         if 'http_address' not in self:
@@ -58,10 +46,7 @@ class Layer(plonetesting.Layer):
         else:
             self.profile = None
 
-        if self.is_remote:
-            self._start_remote()
-        else:
-            self._start_local()
+        self._start_selenium()
         atexit.register(self._stop_selenium)
         speed = os.environ.get('GOCEPT_SELENIUM_SPEED')
         if speed is not None:
@@ -73,33 +58,10 @@ class Layer(plonetesting.Layer):
         self['seleniumrc'].session_id = None
         del self['seleniumrc']
 
-    def _start_local(self):
-        parameters = {}
-        if self.profile:
-            parameters['firefox_profile'] = self.profile
-        ff_binary = os.environ.get('GOCEPT_WEBDRIVER_FF_BINARY')
-        if ff_binary:
-            parameters['firefox_binary'] = FirefoxBinary(ff_binary)
-        module = getattr(selenium.webdriver, self._browser)
-        self['seleniumrc'] = module.webdriver.WebDriver(**parameters)
-
-    def _start_remote(self):
-        parameters = {'desired_capabilities': {'browserName': self._browser}}
-        ff_binary = os.environ.get('GOCEPT_WEBDRIVER_FF_BINARY')
-        if ff_binary:
-            parameters['desired_capabilities']['firefox_binary'] = ff_binary
-        if self.profile:
-            parameters['browser_profile'] = self.profile
-
-        try:
-            self['seleniumrc'] = selenium.webdriver.Remote(
-                'http://%s:%s/wd/hub' % (self._server, self._port),
-                **parameters)
-        except urllib2.URLError, e:
-            raise urllib2.URLError(
-                'Failed to connect to Selenium server at %s:%s,'
-                ' is it running? (%s)'
-                % (self._server, self._port, e))
+    def _start_selenium(self):
+        options = selenium.webdriver.FirefoxOptions()
+        self['seleniumrc'] = selenium.webdriver.Firefox(
+            firefox_profile=self.profile, options=options)
 
     def _stop_selenium(self):
         # Only stop selenium if it is still active.